xen/Coverity: audit of MISSING_BREAK defects
Coverity uses several heuristics to identify when one case statement
legitimately falls through into the next, and a comment as the final item in a
case statement is one heuristic (the assumption being that it is a
justification for the fallthrough).

Use this to perform an audit of defects and hide the legitimate fallthroughs.

No functional change.  All identified fallthroughs are legitimate.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Coverity-IDs: 1055483, 1055484, 1055486 - 1055488, 1055490 - 1055496,
              1055498 - 1055500, 1055501, 1220091
Reviewed-by: Don Koch <dkoch@verizon.com>
Acked-by: Jan Beulich <jbeulich@suse.com>